feat: add display name to SurfServer and update connection address handling
