+++ surf-core-api/surf-core-api-common/src/main/kotlin/dev/slne/surf/core/api/common/SurfCoreApi.kt
@@ -19,6 +19,7 @@ interface SurfCoreApi {
 
     fun getCurrentServerName(): String
     fun getCurrentServerCategory(): String
+    fun getCurrentServerDisplayName(): String
     fun getCurrentServer(): SurfServer
 
     fun getServerByName(name: String): SurfServer?
