+++ surf-core-paper/src/main/kotlin/dev/slne/surf/core/paper/PaperBootstrap.kt
@@ -14,7 +14,6 @@ import dev.slne.surf.core.paper.redis.listener.PaperRedisListener
 import dev.slne.surf.core.paper.teleport.TeleportRedisListener
 import io.papermc.paper.plugin.bootstrap.BootstrapContext
 import io.papermc.paper.plugin.bootstrap.PluginBootstrap
-import java.net.InetSocketAddress
 
 @Suppress("UnstableApiUsage")
 class PaperBootstrap : PluginBootstrap {
@@ -30,14 +29,12 @@ class PaperBootstrap : PluginBootstrap {
 
         val server = SurfServer(
             name = surfServerConfig.serverName,
+            displayName = surfServerConfig.serverDisplayName,
             category = surfServerConfig.serverCategory,
             state = SurfServerState.STARTING,
             type = SurfServerType.SERVER,
             maxPlayers = 0,
-            connectionAddress = InetSocketAddress(
-                surfServerConfig.connectionAddress.host,
-                surfServerConfig.connectionAddress.port
-            )
+            externalConnectionAddress = null
         )
 
         surfEventBus.fire(SurfServerStartEvent(surfServerConfig.serverName))
